refactor : LottoPrice, WinningPrice, WinningResult 클래스 적절한 패키지로 이동
